+++ packages/lumx-angularjs/src/utils/index.js
@@ -1,5 +1,6 @@
 /* eslint-disable import/no-unassigned-import */
 
+import './depth_service';
 import './enter-keydown_directive';
 import './event-scheduler_service';
 import './focus-on-init_directive';