+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.0.2][] - 2021-01-18
+
 ### Fixed
 
 -   Fix angular z-index system (revert to previous system)
@@ -941,3 +943,5 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [1.0.0]: https://github.com/lumapps/design-system/tree/v1.0.0
 [unreleased]: https://github.com/lumapps/design-system/compare/v1.0.1...HEAD
 [1.0.1]: https://github.com/lumapps/design-system/tree/v1.0.1
+[unreleased]: https://github.com/lumapps/design-system/compare/v1.0.2...HEAD
+[1.0.2]: https://github.com/lumapps/design-system/tree/v1.0.2